+++ src/ErrorHandler.php
@@ -23,6 +23,11 @@ final class ErrorHandler
     public const DEFAULT_RESERVED_MEMORY_SIZE = 10240;
 
     /**
+     * The fatal error types that cannot be silenced using the @ operator in PHP 8+.
+     */
+    private const PHP8_UNSILENCEABLE_FATAL_ERRORS = \E_ERROR | \E_PARSE | \E_CORE_ERROR | \E_COMPILE_ERROR | \E_USER_ERROR | \E_RECOVERABLE_ERROR;
+
+    /**
      * @var self|null The current registered handler (this class is a singleton)
      */
     private static $handlerInstance;
@@ -88,7 +93,7 @@ final class ErrorHandler
     private static $reservedMemory;
 
     /**
-     * @var array List of error levels and their description
+     * @var array<string, string> List of error levels and their description
      */
     private const ERROR_LEVELS_DESCRIPTION = [
         \E_DEPRECATED => 'Deprecated',
@@ -265,7 +270,24 @@ final class ErrorHandler
      */
     private function handleError(int $level, string $message, string $file, int $line, ?array $errcontext = []): bool
     {
-        if (0 === (error_reporting() & $level)) {
+        $isSilencedError = 0 === error_reporting();
+
+        if (\PHP_MAJOR_VERSION >= 8) {
+            // Starting from PHP8, when a silenced error occurs the `error_reporting()`
+            // function will return a bitmask of fatal errors that are unsilenceable.
+            // If by subtracting from this value those errors the result is 0, we can
+            // conclude that the error was silenced.
+            $isSilencedError = 0 === (error_reporting() & ~self::PHP8_UNSILENCEABLE_FATAL_ERRORS);
+
+            // However, starting from PHP8 some fatal errors are unsilenceable,
+            // so we have to check for them to avoid reporting any of them as
+            // silenced instead
+            if ($level === (self::PHP8_UNSILENCEABLE_FATAL_ERRORS & $level)) {
+                $isSilencedError = false;
+            }
+        }
+
+        if ($isSilencedError) {
             $errorAsException = new SilencedErrorException(self::ERROR_LEVELS_DESCRIPTION[$level] . ': ' . $message, 0, $level, $file, $line);
         } else {
             $errorAsException = new \ErrorException(self::ERROR_LEVELS_DESCRIPTION[$level] . ': ' . $message, 0, $level, $file, $line);